Allows adding members of any visibility during debugging session.
Due to CLR bug 1094313 the added members can't be referenced from another assembly. For now we report a rude edit if we detect a MemberRef or TypeRef to a symbol added during debugging.
 (changeset 1384081)

+++ Src/Test/PdbUtilities/Pdb/PdbToXml.cs
@@ -47,7 +47,7 @@ namespace Roslyn.Test.PdbUtilities
             this.options = options;
         }
 
-        public unsafe static string DeltaPdbToXml(Stream deltaPdb, IEnumerable<uint> methodTokens)
+        public unsafe static string DeltaPdbToXml(Stream deltaPdb, IEnumerable<int> methodTokens)
         {
             var writer = new StringWriter();
             ToXml(
@@ -55,7 +55,7 @@ namespace Roslyn.Test.PdbUtilities
                 deltaPdb, 
                 metadataReaderOpt: null,
                 options: PdbToXmlOptions.IncludeTokens, 
-                methodHandles: methodTokens.Select(token => (MethodDefinitionHandle)MetadataTokens.Handle((int)token)));
+                methodHandles: methodTokens.Select(token => (MethodDefinitionHandle)MetadataTokens.Handle(token)));
 
             return writer.ToString();
         }
